requirements.txt
apscheduler
beautifulsoup4
Flask
flask_login
flask_mail
flask_wtf
fuzzywuzzy
habanero
pandas
PyJWT
PyJWT
pytest
pytest
python-dotenv
python_dateutil
python_dateutil
PyYAML
PyYAML
Requests
Unidecode
WTForms
xlsxwriter
